feat(ir): scalar evolution detection
Detection of induction variables. Currently, simple binary induction is
supported. Maybe we can utilize union-find and re-associate instructions
to detect more complex patterns.

The union-find data structure is already under `src/collections` folder.

Besides, there are samples in SysY that have SCEVs with modulus, the
interface is kept in the code as a TODO.